Make C API pointers const
This ensures that the underlying object can still be mutated but the
pointer location can not.

This just ensures that calling functions don't have to be overly
defensive about what is happening with pointers passed to the API.
For example the Map pointer can't be changed to a nullptr, or to another
Map object.